Adding --latest-only option
By default, any duplicate packages that are found will be processed, resulting in multiple versions of the same package in the feed. In case you have a build process that retains older versions of VSIX packages, this option can be used to ensure only the latest version of packages with the same ID end up in the feed.