Adding --latest-only option (#15)
* Adding --latest-only option

By default, any duplicate packages that are found will be processed, resulting in multiple versions of the same package in the feed. In case you have a build process that retains older versions of VSIX packages, this option can be used to ensure only the latest version of packages with the same ID end up in the feed.

* Refactoring latest version to be much simpler.

Removing private method and switching to use a one-line LINQ statement.
It's probably slightly less efficient, but I think the simpler code outweighs the minor performance hit.